backend/requirements.txt
# Core framework dependencies
fastapi==0.104.1
uvicorn==0.24.0

# LangChain and AI dependencies
langchain==0.2.16
langchain-chroma==0.1.3
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langgraph==0.2.19
langgraph-checkpoint==1.0.9
langsmith==0.1.116

# OpenAI and ML dependencies
openai==1.44.0
chromadb==0.5.3
chroma-hnswlib==0.7.3
Chroma==0.2.0

# Data processing dependencies
pandas==2.2.2
numpy==1.26.4
pypdf==4.3.1
SQLAlchemy==2.0.34

# Web and API dependencies
Flask==3.0.0
Flask-CORS==4.0.0

# Validation and configuration
pydantic==2.9.0
pydantic_core==2.23.2
PyYAML==6.0.2

# Utility dependencies
pyprojroot==0.3.0
ipykernel==6.29.5
